Committed backend and frontend changes

+++ Backend2/demo/src/main/java/com/example/demo/repository/UserRepository.java
@@ -2,11 +2,12 @@ package com.example.demo.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.web.bind.annotation.CrossOrigin;
 
 import com.example.demo.model.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
 }
